Fix for multiplayer
